% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{circacf}
\alias{circacf}
\title{Compute ACF of circular data}
\usage{
circacf(x, tau_max)
}
\arguments{
\item{x}{The input data.}

\item{tau_max}{The max correlation lag.}
}
\description{
Computes the autocorrelation of circular data,
**defined on the range [-pi, pi], expressed in radians.**
ACF is computed in 'cov'-type.
}
