% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering.compute.neighborhood}
\alias{clustering.compute.neighborhood}
\title{Neighbourhood computation for efficient density-based geometric clustering.}
\usage{
clustering.compute.neighborhood(rc, radius, fe_prefix, nn)
}
\arguments{
\item{rc}{Character, name of the reaction coordinates file.}

\item{radius}{Numeric, radius for density estimation.}

\item{fe_prefix}{Character, prefix of the free energy file (underscore and
radius will be appended). Default is "fe".}

\item{nn}{Character, name of the nearest-neighbours output file.}
}
\description{
TODO: description
}
