% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering.coring}
\alias{clustering.coring}
\title{Variable Dynamic Coring}
\usage{
clustering.coring(traj, traj_cored, states, wsizes, wsize_default = NULL)
}
\arguments{
\item{traj}{Character, name of the file holding the state trajectory.}

\item{traj_cored}{Character, name of the output file holding the cored
state trajectory.}

\item{states}{Numeric, vector of state ids.}

\item{wsizes}{Numeric, vector of window sizes per state, i.e. use a coring
window of size \code{wsizes[i]} for state \code{states[i]}.}

\item{wsize_default}{Numeric, window size that should be used for all states
not in \code{states}.}
}
\description{
A state transition is accepted only if the trajectory spends a minimum
time (given by the window sizes) in the new states. Otherwise the frame is
reassigned to the previous state. The resulting trajectory exhibits clear-cut
state boundaries.
}
