% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering.densityNetwork}
\alias{clustering.densityNetwork}
\title{Construct density network from landscape screening.}
\usage{
clustering.densityNetwork(minpop, clust_prefix = "clust", step = 0.1)
}
\arguments{
\item{minpop}{Numeric, minimum population to count a microstate as node.}

\item{clust_prefix}{Character, prefix of the clustering files obtained by
screening the free energy landscape (se \code\link{clustering.screening}).}

\item{step}{Numeric, step size, i.e. free energy differences between nodes.}
}
\description{
Output files are
\itemize{
\item remapped_clust_<feThreshold>
\item network_nodes.dat, network_links.dat, network_leaves.dat
\item network_end_node_traj.dat
\item network_visualization.html
}
Note that these files will be written to the directory that contains the
<clust_prefix>_<radius> files.
}
\details{
TODO clustering files currently need to be named clust.<radius>
}
