% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering.estimate.radii}
\alias{clustering.estimate.radii}
\title{Neighbourhood populations and free energies for density-based clustering.}
\usage{
clustering.estimate.radii(rc, radii, fe_prefix, pop_prefix)
}
\arguments{
\item{rc}{Character, name of the reaction coordinates file.}

\item{radii}{Numeric vector, radii. The radius defines the size of the
neighbourhood.}

\item{fe_prefix}{Character, free energy output file prefix
(underscore and radius will be appended). Default is "fe".}

\item{pop_prefix}{Character, populations output file prefix
(underscore and radius will be appended). Default is "pop".}
}
\description{
Compute neighbourhood populations and free energies for different radii.
}
\details{
For each frame in the reaction coordinates file \code{rc} and each radius
in \code{radii} the neighbourhood population and corresponding free energies
are computed. The results are written to <pop_prefix>_<radius> and
<fe_prefix>_<radius>, respectively.
}
