% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering.microstates}
\alias{clustering.microstates}
\title{Compute Microstate trajectory.}
\usage{
clustering.microstates(rc, radius, fe_prefix, nn, init, microstates,
  sorted = TRUE)
}
\arguments{
\item{rc}{Character, name of reaction coordinates file.}

\item{radius}{Numeric, density estimation radius.}

\item{fe_prefix}{Character, prefix of the free energy files (underscore and
radius will be appended).}

\item{nn}{Character, name of the nearest-neighbours file.}

\item{init}{Character, name of the file specifying the initial state
definition.}

\item{microstates}{Character, name of the file storing the resulting state
trajectory.}

\item{sorted}{Logical, should microstate names be ordered by descending
populations? (Default: \code{TRUE})}
}
\description{
Assign frames to a distinct microstate using the tree structure obtained by
\link{clustering.densityNetwork}.
The resulting assignment is written to file <microstates>.
}
