% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering.microstates.renamed}
\alias{clustering.microstates.renamed}
\title{Rename microstates.}
\usage{
clustering.microstates.renamed(microstates, output = NULL)
}
\arguments{
\item{microstates}{Character or vector, either filename or microstate
trajectory.}

\item{output}{Character, filename of renamed microstate trajectory.
If \code{NULL}, no output file is created.}
}
\value{
renamed microstate trajectory
}
\description{
Rename microstates from 1 to N, sorted by descending populations.
}
