% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering.mpp}
\alias{clustering.mpp}
\title{MPP dynamic clustering}
\usage{
clustering.mpp(microstates, fe_prefix, radius, lagtime, mpp_prefix, Q = NULL,
  Qmin = 0.01, Qstep = 0.01, Qmax = 1)
}
\arguments{
\item{microstates}{Character, name of the file storing the state trajectory
obtained via geometric clustering.}

\item{fe_prefix}{Character, prefix of the free energy files (underscore and
radius will be appended).}

\item{radius}{Numeric, density estimation radius.}

\item{lagtime}{Numeric, number of frames to be skipped when calculating.
transition probabilites (timescales below this lagtime will be discarded).}

\item{Q}{Numeric, metastability criterion (\eqn{\in (0,1]})}

\item{Qmin}{Numeric, minimum metastability (\eqn{\in (0,1]})}

\item{Qstep}{Numeric, step size}

\item{Qmax}{Numeric, maximum metastability (\eqn{\in (0,1]})}

\item{mpp}{Character, prefix of the MPP population and MPP trajectory files.}
}
\description{
Peform dynamic clustering using the most-probable-path algorithm to assign
geometric microstates to metastable states.
Either specify a single metastability criterion \code{Q} or a range of
metatstability values by specifying \code{Qmin}, \code{Qmax}, \code{Qstep}.
(If \code{Q} is not \code{NULL}, \code{Qmin}, \code{Qmax}, \code{Qstep} will
be ignored.)
}
\examples{
 clustering.mpp(microstates = "microstates",
                fe_prefix   = "freeEnergy",
                radius      = 0.3,
                lagtime     = 50,
                mpp_prefix  = "mpp",
                Q           = 0.3)
 clustering.mpp(microstates = "microstates",
                fe_prefix   = "freeEnergy",
                radius      = 0.3,
                lagtime     = 50,
                mpp_prefix  = "mpp",
                Qmin        = 0.1,
                Qstep       = 0.01,
                Qmax        = 1)
}
