% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering.screening}
\alias{clustering.screening}
\title{Screening (density-based geometric clustering).}
\usage{
clustering.screening(rc, radius, fe_prefix, nn, clust_prefix, min = 0.1,
  max = NULL, step = 0.1)
}
\arguments{
\item{rc}{Character, name of the reaction coordinates file.}

\item{radius}{Numeric, radius for density estimation.}

\item{fe_prefix}{Character, prefix of the free energy file (underscore and
radius will be appended).}

\item{nn}{Character, name of the nearest-neighbours file.}

\item{clust_prefix}{Character, prefix of the clustering output files
(underscore and energy threshold will be appended).}

\item{min}{Numeric, minimum free energy threshold (default: 0.1)}

\item{max}{Numeric, maximum free energy threshold (default: \code{NULL},
i.e. the maximum free energy level in the dataset)}

\item{step}{Numeric, step size by which the free energy threshold
is increased.}
}
\description{
Iterative clustering of frames below a specific free energy level.
The resulting state assignments are written to <clust_prefix>.<threshold>
where a state assignment of 0 indicates a frame with free energy larger
than the cut-off value.
}
\details{
The parameters \code{min}, \code{max}, and \code{step} define which energy
thresholds are considered. The default value for both \code{min} and
\code{step} is 0.1.
If \code{max} is \code{NULL}, the maximum threshold corresponds to the
maximum free energy level in the dataset.
}
