% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering.wtDistributions}
\alias{clustering.wtDistributions}
\title{Waiting Time Distributions}
\usage{
clustering.wtDistributions(traj, wtd_prefix, wsizes)
}
\arguments{
\item{traj}{Character, name of the file holding the state trajectory.}

\item{wtd_prefix}{Character, prefix of the output file holding
the waiting time distribution (window size and state id will be appended).}

\item{wsizes}{Numeric, vector of window size.}
}
\description{
Compute waiting time distributions for the give window sizes.
 The resulting distributions will be written to
 <wtDistribution_prefix>_<window size>_<state id>.
}
