% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastacf}
\alias{fastacf}
\title{Compute ACF using OpenMP}
\usage{
fastacf(x, tau_max)
}
\arguments{
\item{x}{The input data.}

\item{tau_max}{The max correlation lag.}
}
\description{
ACF is computed in 'cov'-type.
}
