% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_coords.R
\name{generate.caDistances}
\alias{generate.caDistances}
\title{C\eqn{\alpha} distances.}
\usage{
generate.caDistances(ref, traj, residue.mindist = 4, residue.maxdist = NULL,
  ignoreCache = FALSE)
}
\arguments{
\item{ref}{Character, name of the PDB file describing the reference structure.}

\item{traj}{Character, name of the XTC file describing the trajectory.}

\item{residue.mindist}{(default: 4)}

\item{residue.maxdist}{(default:\code{NULL})}

\item{ignoreCache}{Logical, if \code{TRUE} recompute even if output file
already exists (default: \code{FALSE}).}
}
\description{
Generate a trajectory with C\eqn{\alpha} distances as coordinates.
}
