% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_coords.R
\name{generate.cossinTransform}
\alias{generate.cossinTransform}
\title{Cosine/Sine Transformation.}
\usage{
generate.cossinTransform(dihedrals, ignoreCache = FALSE)
}
\arguments{
\item{dihedrals}{Character, name of .dih file containing dihedral angles.}

\item{ignoreCache}{Logical, if \code{TRUE} recompute even if output file
already exists (default: \code{FALSE}).}
}
\description{
Generate cos/sin-transformed dihedrals.
}
