% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_coords.R
\name{generate.dihedrals}
\alias{generate.dihedrals}
\title{Compute dihedral angles.}
\usage{
generate.dihedrals(ref, traj, skipCA = NULL, ignoreCache = FALSE)
}
\arguments{
\item{ref}{Character, name of the PDB file describing the reference structure.}

\item{traj}{Character, name of the XTC file describing the trajectory.}

\item{skipCA}{Numeric vector, C\eqn{_\alpha} indices to be skipped in
increasing order. \cr
If \code{NULL} (default) first and last C\eqn{_\alpha} indices are skipped
(incomplete dihedrals at terminal regions).}

\item{ignoreCache}{Logical, if \code{TRUE} recompute even if output file
already exists (default: \code{FALSE}). Generates a warning if cached files
exist.}
}
\description{
Dihedral angles for the given trajectory are computed using gmx.
The result is saved to <traj>.dih. A file <traj>.dih.info is generated
containing additional information (see \code{\link{write.dihedrals.info}}).
}
\details{
If \code{ignoreCache} is set to \code{FALSE}, both output files already exist
and the skipped residuals in the .dih.info match \code{skipCA}, no
computation is performed.
In any other case, the results are recomputed (and potentially existing
output files are overwritten).
}
