% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_coords.R
\name{generate.filteredCoordinates}
\alias{generate.filteredCoordinates}
\title{Filter a data set}
\usage{
generate.filteredCoordinates(coords, filter, states = NULL, output = NULL,
  ignoreCache = FALSE)
}
\arguments{
\item{coords}{Character, name of the coordinates file.}

\item{filter}{Numeric, some integer \eqn{N}.
If \code{states} is \code{NULL} (default), every \eqn{N}-th frame will be selected.
If \code{states} is a filename, N is selected state.}

\item{states}{Character, name of state trajectory file
(default: \code{NULL}, i.e. not used)}

\item{output}{Character, name of the ouput file.
If \code{NULL} (default) name is constructed from the input parameters.}

\item{ignoreCache}{Logical, if \code{TRUE} recompute even if output files
already exists (default: \code{FALSE}).}
}
\description{
Filter a dataset either by selecting every \eqn{N}-th frame or by selecting
all frames corresponding to state \eqn{N}.
}
