% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_coords.R
\name{generate.reactionCoordinates}
\alias{generate.reactionCoordinates}
\title{Generate reaction coordinates.}
\usage{
generate.reactionCoordinates(coords, columns, output, ignoreCache = FALSE)
}
\arguments{
\item{coords}{Character, either single filename or list of filenames of the
coordinate file(s).}

\item{columns}{Numeric, either single vector or list of vectors indicating
the columns to be selected per coordinate file.}

\item{output}{Character, name of the output file.}

\item{ignoreCache}{Logical, if \code{TRUE} recompute even if output files
already exists (default: \code{FALSE}).}
}
\description{
Generate reaction coordinates by selecting columns of the coordinate file(s).
An additional <output>.info file is generated giving additional information
on which columns were selected.
}
