% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{make.readme}
\alias{make.readme}
\title{Create a README file.}
\usage{
make.readme(readme = "README", dirs)
}
\arguments{
\item{readme}{Character, name of the README file.}

\item{dirs}{Character, vector of directories that store the files to be
listed in the README.}
}
\description{
Creates a README file to describe automatically generated files.
Note that the name of the reaction coordinates files needs to contain the
substring 'reaction_coords'.
}
