% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{msg}
\alias{msg}
\title{Warning/Error Message.}
\usage{
msg(type, arg = NULL)
}
\arguments{
\item{type}{Character, type of warning/error}

\item{arg}{Additional information.}
}
\description{
Get the warning/error message for the given type of warning/error.
}
\details{
Possible warning/error types are:
\itemize{
  \item \code{caching} - Cached files exist. \cr
                         \code{arg} should give the function that produced
                         the warning.
  \item \code{noExec}  - Command could not be executed.
  \item \code{format}  - File does not have the expected format.
  \item \code{missingFile} - An input file does not exist.
}
}
