% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msm.R
\name{msm.countMatrix}
\alias{msm.countMatrix}
\title{Transition count matrix.}
\usage{
msm.countMatrix(traj, lag, traj_ids = NULL)
}
\arguments{
\item{traj}{Character or numeric vector, name of trajectory file or vector
describing the state trajectory.}

\item{lag}{Numeric, time lag to use for counting jumps expressed in number
number of frames, traj[i] -> traj[i+lag].}

\item{traj_ids}{Numeric vector, frame ids identifying the trajectory. These
should be set if the state trajectory is actually composed of several,
independently run simulations. If \code{NULL} a single trajectory is assumed.
Must be of same length as the state trajectory.}
}
\description{
Construct a transition count matrix from a trajectory.
Jumps are encoded as jump from row-index to column-index, i.e.
C[1,2] == 10 means 10 jumps from state 1 to state 2, while
C[2,1] == 20 means 20 jumps from state 2 to state 1.
}
