% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msm.R
\name{msm.mfpt}
\alias{msm.mfpt}
\title{Mean first passage times (MFPT)}
\usage{
msm.mfpt(T, initial_state, n_chains, n_steps)
}
\arguments{
\item{T}{Matrix, column-normalized transition matrix encoding the MSM.}

\item{initial_state}{Numeric, state for which MFPTs should be determined.}

\item{n_chains}{Numeric, number of chains to be run
(more chains = better sampling).}

\item{n_steps}{Numeric, number of timesteps. Should be higher then the
slowest timescale of the system in order to sample of MFPTs.}
}
\description{
Compute the Mean first passage times (MFPT) from one state to all other
states.
Computation is done by Markov Chain Monte Carlo sampling of the
given (col-normalized) transition matrix.
}
