% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msm.R
\name{msm.sim}
\alias{msm.sim}
\title{MSM State Trajectory Simulation.}
\usage{
msm.sim(T, initial_state, n_steps)
}
\arguments{
\item{T}{Matrix, column-normalized transition matrix encoding the MSM.}

\item{initial_state}{Numeric, initial state (must be in range
\code{1:nrow(T)}).}

\item{n_steps}{Numeric, number of simulation steps.}
}
\description{
Generate a state trajectory from a Markov State Model. The state trajectory
will have length \code{n_steps+1} (n_steps simulation steps plus the
initial state).
State names are in the range of \code{1:nrow(T)}.
}
