% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msm.R
\name{msm.transitionMatrix}
\alias{msm.transitionMatrix}
\title{Transition Matrix.}
\usage{
msm.transitionMatrix(traj, lag, traj_ids = NULL, row.normalized = FALSE)
}
\arguments{
\item{traj}{Character or numeric vector, name of trajectory file or vector
describing the state trajectory.}

\item{lag}{Numeric, time lag to use for counting jumps expressed in number
number of frames, traj[i] -> traj[i+lag].}

\item{traj_ids}{Numeric vector, frame ids identifying the trajectory. These
should be set if the state trajectory is actually composed of several,
independently run simulations. If \code{NULL} a single trajectory is assumed.
Must be of same length as the state trajectory.}

\item{row.normalized}{Logical, if \code{FALSE} (default) the transition
matrix is column-normalized, i.e. the entry T_ij encodes the probability
of a jump from j to i.
Otherwise, transition matrix will be row-normalized, i.e. the entry T_ij
encodes the probability of a jump from i to j.}
}
\description{
Construct a transition matrix from a trajectory.
}
