% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msm.R
\name{msm.transitionTimes}
\alias{msm.transitionTimes}
\title{Transition times and life times.}
\usage{
msm.transitionTimes(tmat, lag, dt, row.normalized = FALSE)
}
\arguments{
\item{tmat}{Matrix, transition matrix.}

\item{lag}{Numeric, lagtime (expressed in number of frames) that was used
for estimating the transition matrix estimation.}

\item{dt}{Numeric, time step for a single frame. This defines the unit of
time for the output.}

\item{row.normalized}{Logical, Is tmat row-normalized, i.e. transitions are encoded
as  T_ij: transition i -> j (default: FALSE) or
col-normalized (T_ij: j -> i, this is the default)?}

\item{row.normalized}{Logical, if \code{FALSE} (default) the transition
matrix \code{tmat} is column-normalized, i.e. the entry \code{tmat[i,j]}
encodes the probability of a jump from j to i.
Otherwise, transition matrix will be row-normalized, i.e. the entry
\code{tmat[i,j]} encodes the probability of a jump from i to j.}
}
\description{
Compute transition and life times from a transition matrix.
}
