% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt.autocor}
\alias{plt.autocor}
\title{Plot autocorrelation.}
\usage{
plt.autocor(coords, columns = NULL, lag.max = 0.25, circular = F,
  dt = NULL, logy = F)
}
\arguments{
\item{coords}{Character or (list of) data.frame(s), vector of coordinate
filenames or a (list of) data frame(s).
 If multiple filenames/data frames are given, mean autocorrelation and
 variance are plotted.}

\item{columns}{Numeric vector, column indices of observables to compute
ACF for.}

\item{lag.max}{Numeric, maximal lagtime. If < 1 (default: 0.25), this is
interpreted as a ratio to the total length. Otherwise the lagtime is
assumed to be given in number of timesteps/frames.}

\item{circular}{Logical, should ACF be computed for circular variables,
e.g. dihedral angles? (default: \code{FALSE})}

\item{dt}{Numeric, timestep in [ps]. If not \code{NULL}, the time axis
will be scaled accordingly. Otherwise (default) time is expressed in the
number of timesteps.}

\item{logy}{Logical, plot with logarithmic y-scale. Note: If \code{TRUE},
negative values are filtered out!}
}
\description{
For a single coordinates file the autocorrelation as a function of time lag
is plotted for each column.
For multiple coordinate files the mean autocorrelation and variance is
plotted (for each column the mean/variance is compute over all files).
}
