% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt.cumFlucts}
\alias{plt.cumFlucts}
\title{Plot cumulative fluctuations.}
\usage{
plt.cumFlucts(eigenvals, labels)
}
\arguments{
\item{eigenvals}{Character vector, names of .val/.valn files holding the
eigenvalues.}

\item{labels}{Character vector, PCA method descriptions corresponding to the
respective files in \code{eigenvals}.}
}
\description{
Takes all available PCAs and plots their cumulative fluctuations in a single plot.
}
