% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt.matrix}
\alias{plt.matrix}
\title{Matrix plot.}
\usage{
plt.matrix(x, diverge = FALSE, fancy = FALSE, zlim = NULL)
}
\arguments{
\item{x}{Character or matrix object, filename of table-formatted file or a
matrix object.}

\item{diverge}{Logical, should a diverging color palette be used?
(default:\code{FALSE}).}

\item{fancy}{Logical, should fancy plotting with interactive elements be
performed? (default: \code{FALSE}).}

\item{zlim}{Numeric vector, minimum and maximum value for color scale.
If \code{NULL} (default) limits are given by the min/max values of the matrix.}
}
\description{
Plots a matrix as colored tiles.
}
