% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt.mppNetwork}
\alias{plt.mppNetwork}
\title{Plot hierarchical network of MPP lumping}
\usage{
plt.mppNetwork(dirname)
}
\arguments{
\item{dirname}{Directory of MPP run}
}
\description{
Plot hierarchical network of MPP lumping
}
