% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt.pcaOverview}
\alias{plt.pcaOverview}
\title{PCA overview}
\usage{
plt.pcaOverview(coords, pcs, corr = FALSE)
}
\arguments{
\item{coords}{Character, name of coordinates file for which PCA has been
performed.}

\item{pcs}{Numeric vector, PC indices.}

\item{corr}{Logical, if \code{TRUE} use correlation-based PCA (default:
\code{FALSE}).}
}
\description{
Plot 2d-proj, 1d-proj and eigenvector content for given PCA.
}
