% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt.pcaProj}
\alias{plt.pcaProj}
\title{Plot a 2d-projection for a given PCA.}
\usage{
plt.pcaProj(projected, dim1 = 1, dim2 = 2, diverge = FALSE)
}
\arguments{
\item{projected}{Character or data.frame, filename of the .proj file or data
frame containing the projected data.}

\item{dim1}{Numeric, index of PC to be used as first dimension, i.e. column
to select from file or data frame (default: 1).}

\item{dim2}{Numeric, index of PC to be used as second dimension, i.e. column
to select from file or data frame (default: 2).}

\item{diverge}{Logical, should a diverging color scale be used?
(default: \code{FALSE})}
}
\description{
plots a 2d free energy landscape (in kT) for the given PCA projection.
}
