% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt.populations}
\alias{plt.populations}
\title{Plot sorted per-frame populations for given radii.}
\usage{
plt.populations(pops, radii = NULL, logy = TRUE, select = NULL)
}
\arguments{
\item{pops}{Character or data frame, prefix of population files or data frame
with neighbourhood populations per frame where each column  corresponds to a
particular neighbourhood radius.}

\item{radii}{Numeric vector, selection of radii. If \code{NULL} (default),
read all available population files.}

\item{logy}{Logical, plot with logarithmic y-scale.}

\item{select}{Numeric, plot only every n-th data point, i.e. for select=100,
the population of every 100-th frame is plotted.
A large value will improve performance significantly.
If \code{NULL} (default) select is chosen such that approx 1000 data points
are plotted.}
}
\description{
Plot sorted per-frame populations for given radii.
}
