% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt.ramachandran}
\alias{plt.ramachandran}
\title{Ramachandran plot.}
\usage{
plt.ramachandran(resno, dihedrals, reslabel = NULL)
}
\arguments{
\item{resno}{Numeric, residue number.}

\item{dihedrals}{Character or data.frame, name of the .dih file or data.frame
containing the dihedral angles.}

\item{reslabel}{Character, label of the residue (to be used in the title).}
}
\value{
ggplot object
}
\description{
Plot the distribution of \eqn{\phi} and \eqn{\psi} angles for a given residue.
}
