% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt.ramacolor}
\alias{plt.ramacolor}
\title{Ramacolor plots of state geometries.}
\usage{
plt.ramacolor(statetraj, dihedrals)
}
\arguments{
\item{statetraj}{Character or data frame, filename of state trajectory file
or data frame holding the state trajectory.#' .}

\item{dihedrals}{Character or data frame, filename of the .dih file or data
frame holding dihedral angles.}
}
\description{
Color-coded dihedral content per state and residue.
}
\details{
TODO: all residues are read from file and indexed 1:n_residues
}
