% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt.stateTrajComparison}
\alias{plt.stateTrajComparison}
\title{Plot state trajectory comparison.}
\usage{
plt.stateTrajComparison(traj1, traj2)
}
\arguments{
\item{traj1}{Either a vector encoding first state trajectory or
a filename pointing to the trajectory.}

\item{traj2}{Either a vector encoding second state trajectory or
a filename pointing to the trajectory.}
}
\description{
Compare state trajectories based on their overlap of state populations.
Trajectories must be of same length and must have identical state labels.
Colored states are taken from \code{traj1} and separate into black states,
taken from \code{traj2}.
}
