% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plt.wtDistributions}
\alias{plt.wtDistributions}
\title{Plot waiting time distributions.}
\usage{
plt.wtDistributions(wtd, max_frame = NULL, logy = F)
}
\arguments{
\item{wtd}{data.frame, waiting time distributions}

\item{max_frame}{Numeric, limit the x-axis to obtain a nicer plot.}

\item{ploty}{Logical, plot with logarithmic y-scale.}
}
\description{
Plot waiting time distributions.
}
