% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rama2rgb}
\alias{rama2rgb}
\title{Compute RGB color code from sampling of Ramachandran plot}
\usage{
rama2rgb(phis, psis)
}
\arguments{
\item{phis}{Sampling of phi values.}

\item{psis}{Sampling of psi values.}
}
\description{
phi and psi values are defined on range [-180, 180].
}
