% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.caDistances.info}
\alias{read.caDistances.info}
\title{Read C\eqn{_\alpha} information.}
\usage{
read.caDistances.info(caDistInfo)
}
\arguments{
\item{caDistInfo}{Character, name of the .caDist.info file.}
}
\value{
named list with elements \code{ref}, \code{traj}, \code{mindist},
  \code{maxdist}
}
\description{
Read a .caDist.info file as generated by \code{\link{write.caDistances.info}}.
}
