% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.cumFlucts}
\alias{read.cumFlucts}
\title{Cumulative fluctuations.}
\usage{
read.cumFlucts(vals)
}
\arguments{
\item{vals}{Character or numeric vector, either name of the .val/.valn file
or vector of eigenvalues.}
}
\value{
cumulative fluctuations
}
\description{
Cumulative fluctuations.
}
