% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.dihedrals.info}
\alias{read.dihedrals.info}
\title{Read dihedrals information.}
\usage{
read.dihedrals.info(dihedralsInfo)
}
\arguments{
\item{dihedralsInfo}{Character, name of the .dih.info file.}
}
\value{
named list with elements \code{ref}, \code{traj}, \code{nRes},
  \code{skippedResnos}
}
\description{
Reads a .dih.info file as generated by \code{\link{write.dihedrals.info}}.
}
