% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.dihedrals}
\alias{read.dihedrals}
\title{Read dihedrals file.}
\usage{
read.dihedrals(dihedrals, resnos = NULL, dihedralsInfo = NULL)
}
\arguments{
\item{dihedrals}{Character, name of the .dih file containing the dihedral
angles.}

\item{resnos}{Numeric vector, residue numbers for which dihedral angles
should be read. If \code{NULL} (default) read dihedrals for all residues.}

\item{dihedralsInfo}{Character, name of the .dih.info file as generated by
\code{\link{generate.dihedrals}}. If \code{NULL} the filename is assumed to
be <dihedrals>.info}
}
\value{
A data.frame with \eqn{\phi} and \eqn{\psi} angles.
}
\description{
Read a file containing dihedral angles. A .dih.info file as generated by
\code{\link{write.dihedrals.info}} is needed to correctly identify the
corresponding residues.
}
