% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.populations}
\alias{read.populations}
\title{Read population files.}
\usage{
read.populations(pop_prefix, radii = NULL)
}
\arguments{
\item{radii}{Numeric vector, selection of radii. If \code{NULL} (default),
read all available population files.}

\item{prefix_pop}{Character, population files prefix.
File names are assumed to be <prefix_pop>_<radius>.}
}
\value{
data frame where each column stores the neighbourhood population
 per frame for a particular neighbourhood radius.
}
\description{
Read population files with the given prefix and return them as a
data frame.
}
