% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.wtDistributions}
\alias{read.wtDistributions}
\title{Read waiting time distributions.}
\usage{
read.wtDistributions(wtd_prefix, states, wsizes)
}
\arguments{
\item{wtd_prefix}{Character, prefix of the distribution files
(window size and state id will be appended).}

\item{wsizes}{Numeric, vector of coring window sizes.}

\item{state}{Numeric, vector of state ids.}
}
\value{
data frame with columns `frame`, `probability`, `state` and `wsize`.
}
\description{
Read waiting time distribution files <wtd_prefix>_<wsize>_<state> as
generated by \code{\link{clustering.wtDistributions}} for given states
and window sizes.
}
