% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{reverselog_trans}
\alias{reverselog_trans}
\title{Negative Logarithm.}
\usage{
reverselog_trans(base = exp(1))
}
\arguments{
\item{base}{Numeric, base to which the logarithm is computed
(default: \eqn{e})}
}
\description{
Transform \eqn{x} to \eqn{-log(x)} for plotting.
}
