% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_binaries.R
\name{run.cmd}
\alias{run.cmd}
\title{Run single command.}
\usage{
run.cmd(cmd, args, onError = NULL)
}
\arguments{
\item{cmd}{Character, command to run.}

\item{args}{Character vector, arguments to the command.}

\item{onError}{Function, function to be performed in case of an error.}
}
\description{
Runs command with given arguments.
}
\details{
Wrapper for \code{\link{system2}} that stops the execution and throws an
error if the command gives a non-zero exit status.
}
