% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_binaries.R
\name{run.cmds}
\alias{run.cmds}
\title{Run commands.}
\usage{
run.cmds(cmds, onError = NULL)
}
\arguments{
\item{cmds}{Character, string of commands to be run.}

\item{onError}{Function, function to be performed in case of an error.}
}
\description{
Runs the string of commands given by cmds.
}
\details{
Wrapper for \code{\link{system}} that stops the execution and throws an error
if the command gives a non-zero exit status.
Use \code{\link{run.cmd}} for a single command.
}
