% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_coords.R
\name{run.dPCAplus}
\alias{run.dPCAplus}
\title{dPCA+}
\usage{
run.dPCAplus(dihedrals, corr = FALSE, ignoreCache = FALSE)
}
\arguments{
\item{dihedrals}{Character, name of .dih file containing dihedral angles.}

\item{corr}{Logical, if \code{TRUE} use correlation instead of covariance.
Effectively whitens the data before doing analysis (default: \code{FALSE}).}

\item{ignoreCache}{Logical, if \code{TRUE} recompute even if output files
already exist (default: \code{FALSE}).}
}
\description{
Runs a dPCA+ analysis on dihedral angles.
}
