% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_coords.R
\name{run.PCA}
\alias{run.PCA}
\title{PCA}
\usage{
run.PCA(coords, corr = FALSE, ignoreCache = FALSE,
  additionalParams = NULL)
}
\arguments{
\item{coords}{Character, name of file containing the coordinates.}

\item{corr}{Logical, if \code{TRUE} use correlation instead of covariance.
Effectively whitens the data before doing the analysis (default:
\code{FALSE}).}

\item{ignoreCache}{Logical, if \code{TRUE} recompute even if output files
already exist (default: \code{FALSE}).}

\item{additionalParams}{Character vector, additional parameters for 'fastpca'
(default: \code{NULL}).}
}
\description{
Run PCA on the given coordinates using fastpca. \cr
The following files are created:
\itemize{
\item if \code{corr} is \code{FALSE}: <coords>.proj, <coords>.vec,
      <coords>.val, <coords>.cov, <coords>.stats
\item if \code{corr} is \code{TRUE}: <coords>.projn, <coords>.vecn,
      <coords>.valn, <coords>.covn, <coords>.statsn
}
}
