% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_binaries.R
\name{set.binary}
\alias{set.binary}
\title{Set path to binary.}
\usage{
set.binary(name, path)
}
\arguments{
\item{name}{Name of the binary.}

\item{path}{Absolute system path to binary.}
}
\description{
Set path to binary.
}
\examples{
set.binary("bash", "/bin/bash")
}
