% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats.autocor}
\alias{stats.autocor}
\title{Autocorrelation.}
\usage{
stats.autocor(x, lag.max = 0.25, columns = NULL, circular = FALSE,
  circ.units = "deg")
}
\arguments{
\item{x}{Character or data.frame, either a filename or a data frame
providing the data.}

\item{lag.max}{Numeric, maximal correlation lag.
If < 1, expressed in fractions of total length (default: 0.25).}

\item{columns}{Numeric vector, selected observables for which the
autocorrelation should be computed. If \code{NULL} compute all.}

\item{circular}{Logical, should autocorrelation be computed for circular
variables, e.g. dihedral angles? (default: \code{FALSE})}

\item{circ.units}{Character, circular units, one of \code{c("deg", "rad")}}
}
\value{
data.frame where row i corresponds to the autocorrelation for a delay
 of i-1 frames.
}
\description{
Correlation between a signal and itself delayed in time.
}
