% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{streamed_cossin_transform}
\alias{streamed_cossin_transform}
\title{computes a cos/sin transform of angles on the fly}
\usage{
streamed_cossin_transform(fname_in, fname_out, is_deg = TRUE)
}
\arguments{
\item{fname_in}{Input filename with angles.}

\item{fname_out}{Output filename with cos/sin transformed values.}

\item{is_deg}{Angles are expressed in degrees. default: true.}
}
\description{
computes a cos/sin transform of angles on the fly
}
