% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmd_interface.R
\name{vmd.renderPostscript}
\alias{vmd.renderPostscript}
\title{render current view to postscript}
\usage{
vmd.renderPostscript(filename = "vmdscene.ps")
}
\arguments{
\item{filename}{Filename of rendered output}
}
\description{
render current view to postscript
}
