% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmd_interface.R
\name{vmd.setStyle}
\alias{vmd.setStyle}
\title{set rendering style}
\usage{
vmd.setStyle(style, molid = 0)
}
\arguments{
\item{style}{The rendering style. One of: "Licorice"}
}
\description{
set rendering style
}
