% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.caDistances.info}
\alias{write.caDistances.info}
\title{Write C\eqn{_\alpha} infomation.}
\usage{
write.caDistances.info(ref, traj, mindist, maxdist, fname = NULL)
}
\arguments{
\item{ref}{Character, name of the PDB file describing the reference structure.}

\item{traj}{Character, name of the XTC file describing the trajectory.}

\item{fname}{Character, filename that should end in .caDist.info \cr
If \code{NULL} (default) the file is named <traj>.caDist.info}

\item{residue.mindist}{Numeric}

\item{residue.maxdist}{Numeric}
}
\description{
Writes a .caDist.info file with the following format (fields are tab
separated)
\preformatted{
reference <normalized path to .pdb file>
trajectory <normalized path to .xtc file>
mindist <integer>
maxdist <integer>
}
}
