% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.dihedrals.info}
\alias{write.dihedrals.info}
\title{Write dihedrals infomation.}
\usage{
write.dihedrals.info(ref, traj, nRes, skipCA, fname = NULL)
}
\arguments{
\item{ref}{Character, name of the PDB file describing the reference structure.}

\item{traj}{Character, name of the XTC file describing the trajectory.}

\item{nRes}{Numeric, largest residue index}

\item{skipCA}{Numeric vector, C\eqn{_\alpha} indices to be skipped in
increasing order.}

\item{fname}{Character, filename that should end in .dih.info \cr
If \code{NULL} (default) the file is named <traj>.dih.info}
}
\description{
Writes a .dih.info file with the following format (fields are tab separated)
\preformatted{
reference <normalized path to .pdb file>
trajectory <normalized path to .xtc file>r
nResidues <integer>
skippedResidues <space separated integers increasingly ordered>
}
}
